Create edit mode switch button sub-section  for able to switch to different edit mode with description.
