+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Settings/TimeTags/TimeTagCreateConfigSubsection.cs
@@ -3,124 +3,17 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using osu.Framework.Allocation;
-using osu.Framework.Bindables;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Shapes;
-using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Localisation;
 using osu.Game.Graphics;
-using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Settings.Components.Markdown;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States.Modes;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Components.Markdown;
-using osuTK;
 using osuTK.Graphics;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Settings.TimeTags;
 
-public partial class TimeTagCreateConfigSubsection : FillFlowContainer, IHasCurrentValue<CreateTimeTagEditMode>
+public partial class TimeTagCreateConfigSubsection : EditModeSwitchSubsection<CreateTimeTagEditMode>
 {
-    private const int button_vertical_margin = 20;
-    private const int horizontal_padding = 20;
-    private const int corner_radius = 15;
-
-    private readonly EditModeButton[] buttons;
-    private readonly LyricEditorDescriptionTextFlowContainer lyricEditorDescription;
-
-    private readonly BindableWithCurrent<CreateTimeTagEditMode> current = new();
-
-    public Bindable<CreateTimeTagEditMode> Current
-    {
-        get => current.Current;
-        set => current.Current = value;
-    }
-
-    [Resolved]
-    private OsuColour colours { get; set; } = null!;
-
-    private readonly Box background;
-
-    public TimeTagCreateConfigSubsection()
-    {
-        RelativeSizeAxes = Axes.X;
-        AutoSizeAxes = Axes.Y;
-        Spacing = new Vector2(10);
-
-        Children = new Drawable[]
-        {
-            new Container
-            {
-                RelativeSizeAxes = Axes.X,
-                AutoSizeAxes = Axes.Y,
-                Masking = true,
-                CornerRadius = corner_radius,
-                Children = new Drawable[]
-                {
-                    background = new Box
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                    },
-                    new GridContainer
-                    {
-                        RelativeSizeAxes = Axes.X,
-                        AutoSizeAxes = Axes.Y,
-                        RowDimensions = new[]
-                        {
-                            new Dimension(GridSizeMode.AutoSize),
-                        },
-                        Content = new[]
-                        {
-                            buttons = Enum.GetValues<CreateTimeTagEditMode>().Select(x => new EditModeButton(x)
-                            {
-                                Text = getButtonTitle(x),
-                                Margin = new MarginPadding { Vertical = button_vertical_margin },
-                                Padding = new MarginPadding { Horizontal = horizontal_padding },
-                                Action = () => current.Value = x,
-                            }).ToArray(),
-                        },
-                    },
-                },
-            },
-            lyricEditorDescription = new LyricEditorDescriptionTextFlowContainer
-            {
-                RelativeSizeAxes = Axes.X,
-                AutoSizeAxes = Axes.Y,
-                Padding = new MarginPadding { Horizontal = horizontal_padding },
-            },
-        };
-
-        current.BindValueChanged(e =>
-        {
-            Schedule(() =>
-            {
-                updateEditMode(e.NewValue);
-            });
-        }, true);
-    }
-
-    [BackgroundDependencyLoader]
-    private void load(LyricEditorColourProvider colourProvider, ILyricEditorState state)
-    {
-        background.Colour = colourProvider.Background4(state.Mode);
-    }
-
-    private void updateEditMode(CreateTimeTagEditMode mode)
-    {
-        // update button style.
-        foreach (var button in buttons)
-        {
-            bool highLight = EqualityComparer<CreateTimeTagEditMode>.Default.Equals(button.Mode, mode);
-            button.Alpha = highLight ? 0.8f : 0.4f;
-            button.BackgroundColour = getButtonColour(button.Mode, highLight);
-        }
-
-        // update description text.
-        lyricEditorDescription.Description = getDescription(mode);
-    }
-
-    private LocalisableString getButtonTitle(CreateTimeTagEditMode mode)
+    protected override LocalisableString GetButtonTitle(CreateTimeTagEditMode mode)
         => mode switch
         {
             CreateTimeTagEditMode.Create => "Create",
@@ -128,7 +21,7 @@ public partial class TimeTagCreateConfigSubsection : FillFlowContainer, IHasCurr
             _ => throw new InvalidOperationException(nameof(mode)),
         };
 
-    private Color4 getButtonColour(CreateTimeTagEditMode mode, bool active)
+    protected override Color4 GetButtonColour(OsuColour colours, CreateTimeTagEditMode mode, bool active)
         => mode switch
         {
             CreateTimeTagEditMode.Create => active ? colours.Green : colours.GreenDarker,
@@ -136,7 +29,7 @@ public partial class TimeTagCreateConfigSubsection : FillFlowContainer, IHasCurr
             _ => throw new InvalidOperationException(nameof(mode)),
         };
 
-    private DescriptionFormat getDescription(CreateTimeTagEditMode mode) =>
+    protected override DescriptionFormat GetDescription(CreateTimeTagEditMode mode) =>
         mode switch
         {
             CreateTimeTagEditMode.Create => new DescriptionFormat
@@ -200,14 +93,4 @@ public partial class TimeTagCreateConfigSubsection : FillFlowContainer, IHasCurr
             },
             _ => throw new InvalidOperationException(nameof(mode)),
         };
-
-    private partial class EditModeButton : EditorSectionButton
-    {
-        public CreateTimeTagEditMode Mode { get; }
-
-        public EditModeButton(CreateTimeTagEditMode mode)
-        {
-            Mode = mode;
-        }
-    }
 }
